requirements.txt
gepa-ai
openai
python-dotenv
Pillow
scikit-learn
pydantic
PyYAML
litellm
